INTEL_DII/FIXME: drm/i915/rpm: Restrict rpm suspend to client platforms
Regressions on SRIOV tests are observed while enabling runtime
autosuspend , the second tile initializations are throwing
errors.So, for now restrict runtime pm autosuspend to 1T.

v2: On a 1T PVC, rpm suspend is being triggered
when ANR is alive and ANR is blocked,
This is a FIXME, i915 should not be runtime suspended,
if a child device like ANR is alive. this need to
be debugged and a proper solution like making sure
of the presence of ANR and holding rpm wakeref
is under discussion.

As this is a blocker for release, we are providing a
temporary fix to block/forbid rpm suspend for XEHPSDV and
PVC platforms.

Signed-off-by: Tilak Tangudu <tilak.tangudu@intel.com>
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>