INTEL_DII: drm/i915/sriov: Detect broken MMIO BAR
In commit 29b6f88d60dd ("drm/i915: Try to detect sudden loss of
MMIO access") we have added mechanism to detect problems with MMIO
BAR, but if MMIO BAR is already broken while we probe for the VF
capability, since we are not using above mechanism, we may wrongly
claim that we are running on the VF device.

To avoid reporting misleading follow-up errors, check that VF_CAP
register does not have any other bits set.

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>