drm/i915: Skip error capture of completely idle engines
When we initiate error capture for an engine (before an isolated engine
reset), we only capture that engine and its request. However, when we
initiate capture for a GT, we capture all engines that have a request.
Before we determine that the engine does have a request, we may modify
the register state that requires an engine reset to recover. But if the
engine is idle, then we will skip the engine reset (as we cannot acquire
the engine-wakeref inside the reset). We need to keep the
already conditional error capture symmetric with the reset, so only look
for a request along the engine if the engine is awake and able to be
reset. (It is a condition of execution that a request keeps the engine
awake.)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>
Tested-by: Andi Shyti <andi.shyti@linux.intel.com>