drm/i915: Use the full-reset domain upon request
If we need to reset the device, and not a particular engine, we pass
ALL_ENGINES to intel_gt_handle_error. However, we then mask that with
the available engines and try to do each one individually. This loses
the information that we wanted to do a full-reset and the special case
for handling a device reset (e.g. gen6_reset_engines).

Fixes: 873d66fb9b1d ("drm/i915: Trim error mask to known engines")
Reported-by: Daniel Mrzygłód <daniel.t.mrzyglod@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: Daniel Mrzygłód <daniel.t.mrzyglod@intel.com>
Cc: <stable@vger.kernel.org> # v4.18+
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Oak Zeng <Oak.Zeng@intel.com>