INTEL_DII/FIXME: drm/i915/gsc: GSC firmware loading
GSC FW is loaded by submitting a dedicated command via the GSC engine.
The memory area used for loading the FW is then re-purposed as local
memory for the GSC itself, so it needs to be big enough.

The GSC is not reset as part of GT reset, so we only need to load it on
first boot and S3/S4 exit.

FIXME: GSC load takes a lot of time, so we can't do it serially doing
probe/resume and should be moved to a work thread.

v2: Handle already-loaded GSC (Alan)
v3: Setting GSC FW loading buffer to be uncached for MTL and flush the
    write combining buffer before unmapping the object.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>