INTEL_DII: drm/i915/vf: Start migration recovery due to CTB
If VF was migrated with CTB enabled, GuC will mark CTB status
as MIGRATED. Look for this special MIGRATED bit on both write
and read as this is not the indication of CTB corruption.

Within ct_send(), we do spin_lock_irqsave() - so we disable interrupts.
Our post-migration worker cannot be scheduled within that block, as that
would use queue_work(), which does local_irq_save(). Kernel would actually
detect that, issuing "WARNING: inconsistent lock state". and another one
about a deadlock. This is why we do not schedule directly in ct_read()
and ct_write().

Co-authored-by: Tomasz Lis <tomasz.lis@intel.com>
Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>