INTEL_DII/NOT_UPSTREAM: drm/i915/guc: mtl has mmio doorbells
MTL and newer integrated platforms have mmio doorbells, to keep the HW
and SW flows the same as the ones for discrete cards.

We currently only use doorbells in an HW tests, so we only need this to
make sure there are no regressions when we test the HW in emu or at PO,
hence the not-upstream tag.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>