INTEL_DII: drm/i915/dmc: Load DMC on MTL
Add support to load DMC v2.12 on MTL.

This change is specific for DII and is NOT to be upstreamed. The
approach on upstream is to have DMC being loaded with unversioned blob
filenames. As MTL support in DII is temporary and should be dropped
later, to keep it simple, just keep using versioned paths (those
binaries will come from our internal linux-firmware).

v2:
    - Update to use v2.11. (Gustavo)
    - Add note in commit message about this being DII-specific and
      diverging from upstream's approach. (Gustavo)
v3:
    - Update to use v2.12. (Gustavo)
v4:
    - Sync maximum payload size with the equivalent from upstream:
      s/MTL_DMC_MAX_FW_SIZE/XELPDP_DMC_MAX_FW_SIZE/ and use 0x7000
      instead of 0x10000. (Gustavo)

Signed-off-by: Madhumitha Tolakanahalli Pradeep <madhumitha.tolakanahalli.pradeep@intel.com>
Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>