drm/i915/gt: Serialise GRDOM access between multiple engine resets
Mika asked what is stopping parallel engine resets from clobbering the
common registers used to control the GPU across a reset. The short and
long answer was nothing. If we reset two engines in parallel, they would
both try to select a reset bit (and send requests to common registers)
and wait on that register, at the same time. Serialise control of the
reset requests/acks using the uncore->lock, which will also ensure that
no other GT state changes at the same time as the actual reset.

Reported-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: Andi Shyti <andi.shyti@intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>