INTEL_DII: drm/i915/xehpsdv: implement GGTT Wa_14011411649
The WA states to avoid mixing SMEM and LMEM allocations within the same
GGTT cacheline (i.e. 16 PTEs, 64k address space). LMEM allocation already
need to be rounded to 64k both in size and alignment, so the only case
where we can mix the two types is when a new SMEM PTE is inserted in a
CL containing stale LMEM PTEs. To avoid that, we align SMEM allocation
to 64k as well; this wastes a bit of the address space, but it isn't
too bad given that the majority of GGTT-mapped objects are in LMEM.

Note: tested in pre-si only

v2: just align SMEM to 64k (Chris), fix WA number

v3: Apply workaround for all A steppings (Harish)

v4 (MattR): This workaround now applies to all steppings.

Suggested-by: Chris P Wilson <chris.p.wilson@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Chris P Wilson <chris.p.wilson@intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Harish Chegondi <harish.chegondi@intel.com>