drm/i915/gt: Reacquire heartbeat request before validation
Since commit bf9bd6a5128a ("drm/i915/gt: Track the most recent pulse for
the heartbeat"), the heartbeat can be updated as soon as it has been
cleared; that is we can inject new heartbeat while the heartbeat worker
is executing. As such, we can clear a completed heartbeat, and then see
a new outstanding one, but since we keep the original request in a
local, we mistakenly compare against the old (freed) request.

Fixes: bf9bd6a5128a ("drm/i915/gt: Track the most recent pulse for the heartbeat")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: <stable@vger.kernel.org> # v5.11+
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>