drm/i915/selftests: Disable heartbeats during perf selftests
In order to accommodate intel_engine_is_idle() no longer being able to
determine if the engine is idle while holding its wakeref, rearrange the
engine pm to be dropped prior to waiting for the engine to idle. In doing
so, we can take advantage of our selftest helpers that also suspend the
heartbeat during the measurements, removing extra overhead during our
measurements.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>